fix: long ping totals (no overflow on large CIDR); mark best/worst per ping column

+++ src/IpScanner/Core/Scanner/ScanEngine.cs
@@ -82,7 +82,7 @@ public sealed class ScanEngine
         int offlineCount = _devices.Count - online.Count;
         if (analysisPerIp > 0 && offlineCount > 0)
         {
-            Progress.AddSkipped(offlineCount * analysisPerIp);
+            Progress.AddSkipped((long)offlineCount * analysisPerIp);
             Progress.NotifyChanged();
         }
